+++ src/client/theme-default/components/VPNavBarTitle.vue
@@ -39,8 +39,8 @@ const target = computed(() =>
     >
       <slot name="nav-bar-title-before" />
       <VPImage v-if="theme.logo" class="logo" :image="theme.logo" />
-      <template v-if="theme.siteTitle">{{ theme.siteTitle }}</template>
-      <template v-else-if="theme.siteTitle === undefined">{{ site.title }}</template>
+      <template v-if="theme.siteTitle"><span>{{ theme.siteTitle }}</span></template>
+      <template v-else-if="theme.siteTitle === undefined"><span>{{ site.title }}</span></template>
       <slot name="nav-bar-title-after" />
     </a>
   </div>