+++ src/client/theme-default/components/VPNavBarTranslations.vue
@@ -1,5 +1,4 @@
 <script lang="ts" setup>
-import VPIconLanguages from './icons/VPIconLanguages.vue'
 import VPFlyout from './VPFlyout.vue'
 import VPMenuLink from './VPMenuLink.vue'
 import { useData } from '../composables/data'
@@ -13,7 +12,7 @@ const { localeLinks, currentLang } = useLangs({ correspondingLink: true })
   <VPFlyout
     v-if="localeLinks.length && currentLang.label"
     class="VPNavBarTranslations"
-    :icon="VPIconLanguages"
+    icon="vpi-languages"
     :label="theme.langMenuLabel || 'Change language'"
   >
     <div class="items">