+++ src/shared/shared.ts
@@ -23,8 +23,10 @@ export type SidebarItem = DefaultTheme.SidebarItem
 
 export const EXTERNAL_URL_RE = /^(?:[a-z]+:|\/\/)/i
 export const APPEARANCE_KEY = 'vitepress-theme-appearance'
-export const HASH_RE = /#.*$/
-export const EXT_RE = /(index)?\.(md|html)$/
+
+const HASH_RE = /#.*$/
+const HASH_OR_QUERY_RE = /[?#].*$/
+const INDEX_OR_EXT_RE = /(?:(^|\/)index)?\.(?:md|html)$/
 
 export const inBrowser = typeof document !== 'undefined'
 
@@ -67,8 +69,10 @@ export function isActive(
   return true
 }
 
-export function normalize(path: string): string {
-  return decodeURI(path).replace(HASH_RE, '').replace(EXT_RE, '')
+function normalize(path: string): string {
+  return decodeURI(path)
+    .replace(HASH_OR_QUERY_RE, '')
+    .replace(INDEX_OR_EXT_RE, '$1')
 }
 
 export function isExternal(path: string): boolean {
@@ -182,24 +186,29 @@ export function slash(p: string): string {
   return p.replace(/\\/g, '/')
 }
 
-const extraExts =
-  (typeof process === 'object' && process.env.VITE_EXTRA_EXTENSIONS) ||
-  (import.meta as any).env?.VITE_EXTRA_EXTENSIONS ||
-  ''
-
-// md, html? are intentionally omitted, see treatAsHtml
-const KNOWN_EXTENSIONS = new Set(
-  (
-    '3g2,3gp,aac,ai,apng,au,avif,bin,bmp,cer,class,conf,crl,css,csv,dll,doc,' +
-    'eps,epub,exe,gif,gz,ics,ief,jar,jpe,jpeg,jpg,js,json,jsonld,m4a,man,' +
-    'mid,midi,mjs,mov,mp2,mp3,mp4,mpe,mpeg,mpg,mpp,oga,ogg,ogv,ogx,opus,otf,' +
-    'p10,p7c,p7m,p7s,pdf,png,ps,qt,roff,rtf,rtx,ser,svg,t,tif,tiff,tr,ts,' +
-    'tsv,ttf,txt,vtt,wav,weba,webm,webp,woff,woff2,xhtml,xml,yaml,yml,zip' +
-    (extraExts && typeof extraExts === 'string' ? ',' + extraExts : '')
-  ).split(',')
-)
+const KNOWN_EXTENSIONS = new Set()
 
 export function treatAsHtml(filename: string): boolean {
+  if (KNOWN_EXTENSIONS.size === 0) {
+    const extraExts =
+      (typeof process === 'object' && process.env?.VITE_EXTRA_EXTENSIONS) ||
+      (import.meta as any).env?.VITE_EXTRA_EXTENSIONS ||
+      ''
+
+    // md, html? are intentionally omitted
+    ;(
+      '3g2,3gp,aac,ai,apng,au,avif,bin,bmp,cer,class,conf,crl,css,csv,dll,' +
+      'doc,eps,epub,exe,gif,gz,ics,ief,jar,jpe,jpeg,jpg,js,json,jsonld,m4a,' +
+      'man,mid,midi,mjs,mov,mp2,mp3,mp4,mpe,mpeg,mpg,mpp,oga,ogg,ogv,ogx,' +
+      'opus,otf,p10,p7c,p7m,p7s,pdf,png,ps,qt,roff,rtf,rtx,ser,svg,t,tif,' +
+      'tiff,tr,ts,tsv,ttf,txt,vtt,wav,weba,webm,webp,woff,woff2,xhtml,xml,' +
+      'yaml,yml,zip' +
+      (extraExts && typeof extraExts === 'string' ? ',' + extraExts : '')
+    )
+      .split(',')
+      .forEach((ext) => KNOWN_EXTENSIONS.add(ext))
+  }
+
   const ext = filename.split('.').pop()
 
   return ext == null || !KNOWN_EXTENSIONS.has(ext.toLowerCase())